Move network policy config to configs
